feat: Add forum post append feature with race-safe implementation and UI toggle\n\n- Add `forum-post-append-window` and `private-forum-post-append-window` config options\n- Add `append` grant action to permission model\n- Add race-safe conditional SQL UPDATE query for appending\n- Fix ordinary edit authorisation in matchers and helpers to correctly enforce `edit` and `edit-any` grants\n- Update forum thread UI to display 'Append' conditionally\n- Add comprehensive test coverage for successful appends, fallback conditions, and permission checks

+++ core/common/coredata.go
@@ -2263,7 +2263,7 @@ func (cd *CoreData) AttemptAppendForumComment(commenterID, threadID, topicID, la
 		CommentID:        commentID,
 		CommenterID:      commenterID,
 		ForumthreadID:    threadID,
-		AppendWindowMins: appendWindowMins,
+		AppendWindowMins: int64(appendWindowMins),
 		Section:          section,
 		ItemType:         sql.NullString{String: itemType, Valid: true},
 		ItemID:           sql.NullInt32{Int32: itemID, Valid: true},
